Move JobPlacement class into PnpJobProcessor
* Change status parameter type in getJobPlacementById from String to JobPlacement.Status
